Fixed threshold condition and also a refactor: extracted the validation method to the super class and override it when neeeded
